### 2019-07-13 #33 @whedon
[PRE REVIEW]: Probabilistic Biostatistics with Julia
**Submitting author:** @tszanalytics (<a href="http://orcid.org/0000-0002-2469-1796">Jeffrey A Mills</a>)
**Repository:** <a href="https://github.com/tszanalytics/Juliacon2019" target ="_blank">https://github.com/tszanalytics/Juliacon2019</a>

Thanks for submitting your paper to JuliaCon @tszanalytics. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@tszanalytics if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
