### 2019-07-15 #36 @whedon
[REVIEW]: Probabilistic Biostatistics with Julia
**Submitting author:** @tszanalytics (<a href="http://orcid.org/0000-0002-2469-1796">Jeffrey A Mills</a>)
**Repository:** <a href="https://github.com/tszanalytics/Juliacon2019" target ="_blank">https://github.com/tszanalytics/Juliacon2019</a>

~- [ ] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)~

~- [ ] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support~

~~**Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support~~
