### 2019-07-18 #36 @tszanalytics
Just an update:  I am putting together example code to generate all the figures and the table in the paper.  I will post this on the repo soon.  I've add a license.  I have added a table and made some other changes to the paper, and I am still working on editing the paper - I will upload a revised version soon.  I will also add the conference presentation and any related code to the repo once I have that together.
