### 2019-07-18 #36 @bkamins
Thank you for all this. So probably it is best if we wait for the update before reviewing. Right?
Please let us know when you are done. Thank you.

Some minor points while you update the text are (this is based on just skimming over the paper - I will read it in derail after the final version is uploaded):
* could you please increase the font side a bit in Fig. 1, 3, 4, 5 (to match the font size of the text) - if possible (BTW: Fig. 2 is a work of art 👍)
* it would be best to avoid breaking the code between pages (like now between page 3 and 4)
* in the code that "overflows" page end indentations are missing in `begin`-`end` and `for` code blocks
* another minor issue (in all codes) is to use spaces consistently (between operators, after commas etc.)
* Citations in IEEE style are usually given before a comma, see http://ieeeauthorcenter.ieee.org/wp-content/uploads/IEEE_Style_Manual.pdf page 12, for a reference
* It would be great if you added a one paragraph on timings of the computations (you only say that it is expensive computationally at the end of the paper now) - sorry if I missed this and you already do it (I just only glanced over the paper now)
