### 2019-07-18 #36 @tszanalytics
Yes, waiting for the update before reviewing makes sense.  I will incorporate all your suggestions above, thanks for these!
