### 2019-07-22 #36 @tszanalytics
Yep, working on my presentation right now!  Will update this after the presentation.
