### 2019-07-24 #36 @bkamins
Some small code review comments:
* n your text `Array{Real}` can be probably replaced by `Array{Float64}` and the code shall run faster.
* you define `wt_draws` twice but do not define `wp_draws` below in the code and probably you do not need to call `Array` when you extract values from `ct`

(I can check it if we have reproducible codes)
