### 2019-07-24 1e97daf
RNG example in presentation
