### 2019-07-26 #36 @tszanalytics
Thanks Bogumil!  I had spotted the wt_draws error but not the Real.  Also, thanks for answering Stefan K's question more clearly than I did at my presentation - I chatted with him later and he said you had explained it to him.

At the airport in Baltimore waiting for my flight right now.  Wish I could have stayed for the hackathon.  Was a fantastic conference.  Hope to chat with you at JuliaCon 2020!
