### 2019-07-26 #36 @bkamins
@tszanalytics - thank you for a very nice presentation (and I did not think that you knew that it was me who made the remark - but it turns out you did 😄).

I am obsessed with the difference in interpretation of hypothesis test results in frequentist and Bayesian inference (and actually I think that Bayesian interpretation is more intuitive for most people) and your work fits perfectly into this picture.
