### 2019-07-26 #36 @tszanalytics
@bkamins - Actually I didn't know it was you at the time, I only found out when I chatted with Stefan later.  I wish I had known so I could have chatted with you.  We must get together at the next time.  My flight delayed so more time to work on the proceedings revisions 😄
