### 2019-08-06 7adae5c
Delete paper.tex
