### 2019-08-06 4e004d5
Delete header.tex
