+++ paper/header.tex
@@ -1,10 +1,0 @@
-% **************GENERATED FILE, DO NOT EDIT**************
-
-\title{Probabilistic Biostatistics with Julia}
-
-\author[1]{Jeffrey A. Mills}
-\author[1, 2]{Jeffrey R. Strawn}
-\affil[1]{University of Cincinnati}
-\affil[2]{Cincinnati Children's Hospital}
-
-\keywords{Julia, Bayesian methods, Meta-analyses, Biostatistics, Randomized Controlled Trial}
