+++ paper/paper.tex
@@ -1,0 +1,228 @@
+% JuliaCon proceedings template
+\documentclass{juliacon}
+\usepackage{amsmath}
+\setcounter{page}{1}
+
+\begin{document}
+
+\input{header}
+
+\maketitle
+
+\begin{abstract}
+
+Medical research requires statistical tools that are both sophisticated and powerful enough to address complex inferential problems, as well as intuitive and user-friendly enough to not require advanced statistical and programming expertise.  Combining Julia with the Bayesian MCMC machinery addresses this need.
+\vskip 6pt
+Examples of Bayesian probabilistic biostatistics in Julia, including Bayesian adaptive trial design, sequential analysis of randomized controlled trials, and Bayesian hierarchical modeling for a meta-analysis are presented.  Further, these examples illustrate the application of a sophisticatedly simple approach to statistical analysis and modeling for clinical research.
+
+\end{abstract}
+
+\section{Introduction}
+
+Julia solves the “two-language problem”: it is both fast and efficient (performance), and easy to use (user friendliness).\cite{bezanson2017julia}  Additionally, using the Bayesian MCMC machinery in Julia solves “the two field problem”: clinical researchers often need expertise in both medicine and in statistical computing.  
+\vskip 6pt
+As those conducting and funding clinical randomized controlled trials (RCTs) recognize the high costs of these studies (e.g., medication expense, time required, and potential exposure of patients to ineffective treatments), there has been greater enthusiasm for (1) improving statistical analytic methods for RCTs, and 2) using evidence-based methods to examine existing naturalistically collected clinical data to inform clinical practice without the need for RCTs.  These approaches require far greater statistical and programming knowledge and sophistication from users.  Thus, there is an urgent need to provide statistical tools to clinician-researchers that are intuitive and easy to use, yet sophisticated and powerful enough “under the hood” to answer questions that simpler methods cannot.
+\vskip 6pt
+The Bayesian machinery of posterior simulation and Markov chain Monte Carlo (MCMC) methods together with Julia offer a solution to this “two-field problem”. This enables exact small sample inference and hypothesis testing for complex models without requiring the restrictive assumptions necessary to obtain analytical tractability (performance), and facilitates the analysis of complex models with basic statistical concepts: frequency distributions, density plots, means, medians, modes, standard deviations, quantiles, and posterior density ratios (user friendliness).\cite{Mills2019} 
+\vskip 6pt
[…]
