+++ juliacon_proceedings2019_code.jl
@@ -1,0 +1,174 @@
+using KernelDensity, CSV, Turing, MCMCChains, Distributions, Plots, StatsPlots
+
+"""
+PDR odds and p-value computation functions from BayesTesting
+(including here to reduce package dependency)
+x = MC sample, h0 = null hypothesis value
+usage: odds, p_val, pval_2 = post_odds_pval(x, h0 = -2.3)
+"""
+function post_odds_pval(x; h0 = 0.0)
+    d = kde(x)
+    p,ind = findmax(d.density)
+    ph0 = ifelse(pdf(d,h0) == 0.0,0.0000001,pdf(d,h0))
+    odds = pdf(d,d.x[ind])/ph0
+    p_val = length(x[x .<= h0])
+    p_val2 = length(x[x .>= h0])
+    if p_val <= p_val2
+        p_value = p_val/length(x)
+    else
+        p_value = p_val2/length(x)
+    end
+    p_value_2tail = 2*p_value
+    return odds, p_value, p_value_2tail
+end
+
+# CAMS data for Treatment and Placebo
+df1 = CSV.read("jcp1.csv")
+df4 = CSV.read("jcp2.csv")
+wk121 = df1.wk121
+wk124 = df4.wk124
+
+# number of posterior simulation draws
+M = 100000
+### Proceedings Figure 1 and Table 1:
+plt1 = plot()
+println("n    odds   prob<0     p-value")
+for i = 6:6:66
+    q = i*2      #### need 2:1 ratio of treated to PCBO
+    m1 = mean(wk121[1:q]); sd1 = std(wk121[1:q])    # group 1 Treatment
+    m4 = mean(wk124[1:i]); sd4 = std(wk124[1:i])    # group 4 Placebo
+    # Distn. of means with draws from analytical posterior t-distributiion
+    # Can use MCMC instead here from conditional posteriors,
+    # equations (3) and (4).
+    t1draw = m1 .+ (sd1/sqrt(q)).*rand(TDist(q-1),M)
+    t4draw = m4 .+ (sd4/sqrt(i)).*rand(TDist(i-1),M)
+    dift1t4 = t1draw - t4draw      # Treatment vs. Placebo
+#    @show(round.(post_odds_pval(dift1t4),digits=4))
+    od, p_val, pval_2 = post_odds_pval(dift1t4)
+    odds = round(od,digits=2)
+    pval2 = round(pval_2,digits=4)
+    pval = round(p_val,digits=4)
+    println(3*i,"   ",odds,"   ",pval,"    ",pval2)
+
+    if i == 6
+        plt1 = plot(dift1t4, st=:density, alpha=(i*0.02*3/2),linewidth=2, xlims=[-12.0,5.0],ylims=[0.0,0.4],label="n = $(3*i)", xlabel="PARS difference",legend=:topleft)
+        vline!([0.0],linewidth=2,linecolor=:black,label="")
+    elseif i < 66
+        plot!(dift1t4, st=:density, alpha=(i*0.02*3/2), linewidth=2,label="n = $(3*i)")
+    else
+        plot!(dift1t4, st=:density, alpha=1.2, linewidth=2,linecolor=:black,label="n = $(3*i)")
+    end
+end
+current()
+#savefig("pars_difference_sequential.png")
+
+### BHM Example ###
+# Turing.jl BHM for binomial trials
+# model specification
+@model binomial_trials(s,n,::Type{TV}=Vector{Float64}) where {TV} = begin
[…]
