### 2019-08-07 #36 @tszanalytics
@bkamins   Ok, recompiled the pdf here.  Thanks for pointing me in the right direction.
