### 2019-08-10 #36 @bkamins
@tszanalytics

I think your work is an interesting case study. Here are my comments to your submission (please treat them as suggestions and decide what you think should be incorporated).

### Formal requirements of the review process (boxes checked-off above)
* **Installation instructions:**: please give information about Julia and package versions used in your work or add Project.toml that specifies this
* **Community guidelines:**: please add information whom to contact in case there are bugs/problems in your code
* **Page limit:**: last page has only a few lines; usually it is recommended to avoid such layout in publications; you can consider rectifying it (while keeping in line with the page limit specified by the proceedings guidelines)

### Code related
* `rng_example.jl`:
    * it is typically recommended to add a space after a comma in the code (also in other files)
    * missing `using Random` on top of the file
    * missing `using StatsPlots` on top of the file
    * probably last line `# savefig("approaching_certainty_03.png")` should be removed from the file as it is commented out
* `logistic_trajectory_by_week_example.jl`
    * `https://github.com/tszanalytics/BayesTesting.jl` is not a registered package; can you please add instructions how it should be loaded by the user (in particular which commit should be used as a reference); also please place all `using` statements at the top of the file (now they are in several places and there are duplicates)
    * there are commented-out lines of code in the file - are they needed?
    * in line 121 you can consider fixing the indentation of the code
* `juliacon_proceedings2019_code.jl`
    * similar comments as for `logistic_trajectory_by_week_example.jl`
    * additionally you can consider breaking very long lines into several lines, like line 132

### Paper related

A major comment is to give more discussion of your BayesTesting.jl package (design goals, functionality etc. - see questions for reviewers to see what kind of information is expected). Of course please try to keep in the page limit of the proceedings when doing this (this is a soft suggestion).
