### 2019-08-11 #36 @bkamins
+ some comments to the PDF in the attached file.
[10.21105.jcon.00036.pdf](https://github.com/JuliaCon/proceedings-review/files/3490508/10.21105.jcon.00036.pdf)
