### 2019-08-20 #36 @matbesancon
@tszanalytics did you finish the corrections from @bkamins ?
