### 2019-08-20 #36 @kescobo
I'm a bit confused about whether I'm supposed to be evaluating just the code in the Proceedings repo, `BayesTesting.jl` or both. This review assumes both, please let me know if I'm mistaken.

## Overall

This manuscript presents a justification for and examples of using Bayesian inference in biomedical research, specifically outlining a method using julia to generate "Bayesian p-values" as a bridge from the current frequentist methods typically employed in biological contexts. As a biologist that's just starting to learn about Bayesian methods, I'm excited to have this resource and to have a force in the julia community pushing towards more robust statistical methods.

The manuscript could benefit from some minor updates to present a consistent voice and to provide additional explanation / context in the form of citations for a number of claims. The accompanying code should have substantially more documentation / clarity, both in terms of content and presentation. That said, this effort is definitely worth publishing with some modification.

### Manuscript

Comments on the manuscript can be found here:
[10.21105.jcon.00036_KSB_comments.pdf](https://github.com/JuliaCon/proceedings-review/files/3522868/10.21105.jcon.00036_KSB_comments.pdf)

### Code

I largely agree with @bkamins ' comments here regarding the need for additional documentation (in a perfect world, there would be runable jupyter notebooks, or you could use something like `Literate.jl`) and tests.

### Other Minor Comments

- There's a "Third Author" in the authors list. I'm guessing that's part of the template slipping through?
- https://github.com/tszanalytics/BayesTesting.jl/issues/11
- Is the raw data for the analyses you perform in the paper accessible? Would be great to have a `DataDeps` annotation for it so it's easily acquirable to reproduce your analyses.
