### 2019-08-21 #36 @matbesancon
Indeed that's both, thanks
