### 2019-08-21 #36 @tszanalytics
Thanks for the comments all.  I have been a bit busy the last week or so, but now this is top of my list.  I will get to work revising the paper and the code, etc. to address the comments.  Adding a jupyter notebook sounds like a good idea also, so I will have a go at putting one together.
