+++ README.md
@@ -1,2 +1,2 @@
 # Juliacon2019
-Paper, presentation and code for JuliaCon 2019 
+Paper, presentation and code (including jupyter notebooks) for JuliaCon 2019 
