### 2019-08-31 67e3509
Delete rng_example.jl
