### 2019-08-31 351aaa0
Delete juliacon_proceedings2019_code.jl
