### 2019-09-05 #36 @matbesancon
ping @tszanalytics, did you have time to work on the corrections?
