### 2019-09-12 fc08373
Delete paper.tex
