+++ paper/ref.bib
@@ -1,112 +1,0 @@
-@article{bezanson2017julia,
-	title={Julia: A fresh approach to numerical computing},
-	author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-	journal={SIAM review},
-	volume={59},
-	number={1},
-	pages={65--98},
-	year={2017},
-	publisher={SIAM}
-}
-@article{Strawn2017,
-author = {Strawn, J.R. and Dobson, E.T. and Mills, J.A. and Cornwall, G.J. and Sakolsky, D. and Birmaher, B. and Compton, S.N. and Piacentini, J. and McCracken, J.T. and Ginsburg, G.S. and Kendall, P.C. and Walkup, J.T. and Albano, A.M. and Rynn, M.A.},
-doi = {10.1089/cap.2016.0198},
-issn = {15578992},
-journal = {Journal of Child and Adolescent Psychopharmacology},
-keywords = {Bayesian,antidepressant,anxiety,clinical trial,response},
-number = {6},
-title = {{Placebo Response in Pediatric Anxiety Disorders: Results from the Child/Adolescent Anxiety Multimodal Study}},
-volume = {27},
-year = {2017}
-}
-@article{Strawn2018,
-author = {Strawn, Jeffrey R and Mills, Jeffrey A and Sauley, Beau A and Welge, Jeffrey A},
-doi = {10.1016/j.jaac.2018.01.015},
-issn = {1527-5418},
-journal = {Journal of the American Academy of Child and Adolescent Psychiatry},
-keywords = {generalized anxiety disorder (GAD),selective serotonin reuptake inhibitor (SSRI, SRI),selective serotonin−norepinephrine reuptake inhibitor (SSNRI, SNRI),separation anxiety disorder (SAD),social phobia (SoP)},
-number = {4},
-pages = {235--244},
-pmid = {29588049},
-title = {{The Impact of Antidepressant Dose and Class on Treatment Response in Pediatric Anxiety Disorders: A Meta-Analysis.}},
-url = {https://linkinghub.elsevier.com/retrieve/pii/S0890856718300534 http://www.ncbi.nlm.nih.gov/pubmed/29588049 http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=PMC5877120},
-volume = {57},
-year = {2018}
-}
-@article{Strawn2018a,
-author = {Strawn, J.R. and Mills, J.A. and Cornwall, G.J. and Mossman, S.A. and Varney, S.T. and Keeshin, B.R. and Croarkin, P.E.},
-doi = {10.1089/cap.2017.0060},
-issn = {15578992},
-journal = {Journal of Child and Adolescent Psychopharmacology},
-keywords = {anxiety,anxiolytic,clinical trial,generalized anxiety disorder (GAD)},
-number = {1},
-title = {{Buspirone in Children and Adolescents with Anxiety: A Review and Bayesian Analysis of Abandoned Randomized Controlled Trials}},
-volume = {28},
-year = {2018}
-}
-@article{Strawn2019,
-	author = {Strawn, Jeffrey R and Mills, Jeffrey A and Croarkin, Paul E},
-	doi = {10.1089/cap.2018.0145},
-	file = {:C$\backslash$:/Users/millsjf/Box/Vita-pubs/jcap{\_}Feb{\_}2019.pdf:pdf},
-	issn = {1557-8992 (Electronic)},
-	journal = {Journal of child and adolescent psychopharmacology},
-	keywords = {depression,depression,major depressive disorder,paroxetine,se,fluoxetine,major depressive disorder,paroxetine,sertraline},
-	number = {4},
-	pages = {250--255},
-	pmid = {30810350},
[…]
