+++ paper/header.tex
@@ -2,9 +2,10 @@
 
 \title{Probabilistic Biostatistics with Julia}
 
-\author[1]{Jeffrey A. Mills}
+\author[1,*]{Jeffrey A. Mills}
 \author[1, 2]{Jeffrey R. Strawn}
 \affil[1]{University of Cincinnati}
 \affil[2]{Cincinnati Children's Hospital}
+\affil[*]{jeffrey.mills@uc.edu}
 
 \keywords{Julia, Bayesian methods, Meta-analyses, Biostatistics, Randomized Controlled Trial, Probabilistic Statistics}
