### 2019-09-12 d5ad2ea
Update paper.tex
