### 2019-09-12 cf7dc16
Update paper.tex
