### 2019-09-12 c7800f5
Delete paper.tex
