### 2019-09-27 fef0bf6
Delete paper.tex
