### 2019-09-28 #36 @tszanalytics
@matbesancon
@bkamins
@kescobo

Thanks everyone for all the great comments and suggestions.  I think the paper is greatly improved as a result.  I believe everything on the github repo is now up to date and all the changes are incorporated there.  Here's a summary of many of the revisions in response to the reviewer suggestions:

•	Changed title.
•	Added email addresses.
•	Put citations before periods and commas.
•	Fixed typos in equations (3) and (4) and added cite to Gelman (2004).
•	Tried changing from Real to Float64, but for this model it turns out it has to be Real because of
use of auto-differentation.
•	Reworded 1st sentence in intro. to make it clearer.
•	Reworded 3rd paragraph so it makes more sense and to remove “clean code”.
•	Provided a citation to support the “iron law” and “institutional inertia” statements.
•	Provided a citation for the Behrens-Fisher problem.
•	Edited references section to correct several typos, etc.
•	Fixed typos in the figures.
•	Increased the font size for the figures to the extent possible (I found this to be particularly difficult!).
•	Reduced the size of Fig. 3 by only including two sets of time plot + density plot, and explained its purpose further. This allowed the addition of a new figure (Fig. 5) that provides an example of posterior density plots for treatment, placebo and difference.
•	Reduced the length of the paper so that it fits on 5 pages.

Please let me know if I've missed anything and I will endeavor to fix it.

Thanks,
Jeff
