### 2019-10-01 #36 @kescobo
@tszanalytics Thanks for the edits. I am largely satisfied with the changes. There are just a few outstanding questions/concerns.

1. I missed this on my first go-through, but it seems like the citation numbers are out of order (eg. Citation 3 comes first, followed by citation 8 etc). I haven't used latex that much, but I thought that citations were put in order automatically. Is something interfering with that? If this is not expected/accepted behavior, I think the citations should be in order.
1. @matbesancon suggested that we should be evaluating both BayesTesting.jl and the content of the manuscript, but @tszanalytics says otherwise. Can we get clarification on that? I haven't thoroughly evaluated the package, and it should be made clear whether that's part of the review process.
1. I am happy to see the notebooks included for analysis. Can you please provide a `Project.toml` and `Manifest.toml` for (just one that includes all the packages necessary for both should be fine). This will be useful for future readers so they can get the exact versions of the packages you use. Also would be good to include the julia version number you're using I think.
