### 2019-10-01 #36 @matbesancon
For the second point, if the manuscript is about and uses BayesTesting.jl, it should be reviewed with it. Otherwise if the software is contained in the repo of the paper, it should be as suggested versioned & documented
