### 2019-10-04 #36 @tszanalytics
1.  I'm not sure about the citations.  It appears to put them in alphabetical order by first author name.  I am using
\bibliographystyle{juliacon}
\bibliography{ref}
as the bibliography style as per instructions, so that must be what is wanted for the proceedings?

2.  My presentation and this paper is not about the package and only uses the package somewhat tangentially.  I presented a preliminary version of the package at JuliaCon 2018, and plan on updating and registering it to provide a more complete package by next year, so I don't think the package itself is ready for evaluation.  The paper and presentation is about application of Julia to analysis of RCTs and meta analysis.

3.  Unfortunately I must admit that I don't really understand toml files and there usage as yet.  I will have to read up on this.  I believe I would have to specify a particular project to enable generation of toml files that contain only the packages for that project?  I'm not sure how to do this yet.
I included the Julia version and computer specifics in the paper on page 3 (at end of right column).

Thanks!
