### 2019-10-04 #36 @bkamins
Ad. 3 - this is very easy:

1. start a terminal in the directory of your project and run a Julia session
2. switch to package manager mode
3. write `activate .`
4. wtire `add WHATEVER_PACKAGES_YOU_REQUIRE_FOR_THE_CODES_TO_WORK`
5. exit Julia

In the directory of your project you will have the files @kescobo mentions.
