### 2019-10-05 #36 @kescobo
1. I agree. If this is the behavior of the styles provided, leave as is
1. Once (3) is done, Someone should be able to clone this repository, instantiate it, and then run the notebooks. When that's ready, let me know and I'll give it a shot.
1. Follow Bogumił's instructions and you should be good
