### 2019-10-16 #36 @tszanalytics
Yes!  Apologies for the delay.  I have been at a conference last week and
headed right back out to another one this week (currently in Chicago) -
everything happens at once!  I should be able to get to this within the
next few days.  Thanks to @bkamins for the clear guidance - it looks very
straightforward with that help.  I will upload the toml files asap.
Thanks for you patience.

Best,
Jeff

On Wed, Oct 16, 2019 at 9:20 AM Mathieu Besançon <notifications@github.com>
wrote:

> thanks both for the review @tszanalytics <https://github.com/tszanalytics>
> could you confirm us when the changes are made?
>
