### 2019-10-29 d094f4f
added TOML files for notebooks, julia 1.2.0
