### 2019-10-29 #36 @tszanalytics
@matbesancon

Ok, the TOML files and updated versions of the notebooks for Julia 1.2.0 are uploaded, so it all should be in order now.

Again, apologies for the long delay.  I made the mistake of upgrading to julia 1.2 which created some problems with Turing, but everything is sorted now.

Thanks for your patience!
Jeff
