### 2019-11-03 #36 @bkamins
I have updated the ticks in the original comment of this topic. I think the one minor thing that could be added (it is not blocking) is that README.md in https://github.com/tszanalytics/Juliacon2019 could have a bit more explanation what files constitute what part of the repository and possibly how they are intended to be used (as it contains: presentation, paper, notebooks and Julia files).
