### 2019-11-07 #36 @kescobo
Thanks for the ping. Just cloned and instantiated the repo - looks great! Boxes checked, thanks for your efforts @tszanalytics, and @matbesancon / @bkamins for helping with review :-) 👍
