### 2020-06-08 #36 @matbesancon
@tszanalytics sorry for the huge delay we had on the proceedings. A minor last thing before accepting, could you add the missing DOI pointed above to the references?
