### 2020-08-12 #36 @matbesancon
bump @tszanalytics
