### 2021-02-18 #36 @kescobo
Looks like there hasn't been any activity on that account since October of last year, and before that, February. Maybe worth reaching out by email (tszanalytics@gmail.com is what's associated with the account) or through other channels? I found his faculty profile pretty easily but not sure it's kosher to post that here.
